Very basic idea of session handling
Just changes the navigation bar based on whether the user is logged in or not. More needs to be done in terms of using authorisation context to allow or disallow access to certain pages. Future work could include an admin context for allowing administrators extra functionality but that's lower priority.